ODERNNDynamics

Documentation for ODERNNDynamics.

NNDynamics.jl learns a model of the environment. It works with the following environments:

Pendulum

Acrobot

Lunar Lander

Installation

In the julia REPL, run

using Pkg
Pkg.add(url="https://github.com/SvenDuve/ODERNNDynamics.jl")

This package is not for general use. To be fully functional, it requires the following packages:

RLTypes

Conda

PyCall

Gymnasium, this is a slightly adapted version. The original package can be found here

Run the following:

using Pkg
Pkg.add(url="https://github.com/SvenDuve/RLTypes.jl")
Pkg.add("Conda")
Pkg.add("PyCall")

Then again within julia,

using Conda
using PyCall
Conda.pip("install", "git+https://github.com/SvenDuve/Gymnasium")

Bring package into scope with

using ODERNNDynamics

Example usage

Once the above is installed, the following code can be used to train an agent on data collected form 100 episodes of the Pendulum environment:

using ODERNNDynamics
using RLTypes
model(Acrobot(), ModelParameter(collect_train=100, collect_test=10))

For further parameters, see the code.

This function returns a data structure containing a trained model.

ODERNNDynamics.ODE_RNN

ODERNNDynamics.ODE_RNN

ODERNNDynamics.accuracy

ODERNNDynamics.modelEnv

ODERNNDynamics.modelEnv

ODERNNDynamics.train_step!

ODERNNDynamics.train_step!

ODERNNDynamics.ODE_RNN — Method

(m::ODE_RNN)(timestamps, datapoints, hidden)

Returns the hidden state of the ODE_RNN model at the given timestamps and datapoints.

ODERNNDynamics.ODE_RNN — Method

ODE_RNN(input_size::Int, hidden_size::Int)

Constructs an ODE_RNN model with the given input and hidden size.

ODERNNDynamics.accuracy — Method

accuracy(y_true, y_pred, tolerance)

Returns the accuracy of the prediction ypred compared to the true value ytrue.

ODERNNDynamics.modelEnv — Method

modelEnv(environment::ContinuousEnvironment, modelParams::ModelParameter)

Main algorithm to train a continuous action model.

ODERNNDynamics.modelEnv — Method

modelEnv(environment::DiscreteEnvironment, modelParams::ModelParameter)

Main algorithm to train a discrete action model.

ODERNNDynamics.train_step! — Method

train_step!(environment::ContinuousEnvironment, S, A, R, S´, T, hidden, timestamps, fθ, model_opt)

Performs a training step on the given continuous action model with gradient descent.

ODERNNDynamics.train_step! — Method

train_step!(environment::DiscreteEnvironment, S, A, R, S´, T, hidden, timestamps, fθ, model_opt, ep::EnvParameter)

Performs a training step on the given discrete action model with gradient descent.